Move package extraction functions to utils
They are now also covered with unit tests.
